requirements.txt
Flask
google-cloud-texttospeech
PyMuPDF
